requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cffi==1.15.0
colorama==0.4.4
coverage==6.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.2
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
entrypoints==0.3
et-xmlfile==1.1.0
Faker==13.2.0
filelock==3.6.0
future==0.18.2
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
nose==1.3.7
nose2==0.10.0
notebook==6.4.6
openpyxl==3.0.9
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.12.48
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.6
pypiwin32==223
pyrsistent==0.18.0
python-dateutil==2.8.2
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
uncertainties==3.1.6
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
